openssl: Fix crash in SHAKE128/256 implementation
OpenSSL currently doesn't support squeezing bytes out of an XOF multiple
times.  Unfortunately, EVP_DigestFinalXOF() completely resets the context
and later calls not simply fail, they cause a null-pointer dereference in
libcrypto.  This fixes the crash but is not fully compatible with our
API.

There is an open issue and PR that might add a function that allows
squeezing more data from an XOF in a future version of OpenSSL.